Rename library to openscap_results_parser
- openscap_parser has a connotation that this library is performing as a SCAP validator, content scanner,
  and/or check engine. This PR removes that incorrect connotation.
