+++ Gemfile
@@ -1,4 +1,4 @@
 source "https://rubygems.org"
 
-# Specify your gem's dependencies in openscap_parser.gemspec
+# Specify your gem's dependencies in openscap_results_parser.gemspec
 gemspec
