+++ README.md
@@ -1,13 +1,13 @@
-# OpenscapParser
+# Openscap Results Parser
 
-Welcome to your new gem! In this directory, you'll find the files you need to be able to package up your Ruby library into a gem. Put your Ruby code in the file `lib/openscap_parser`. To experiment with that code, run `bin/console` for an interactive prompt.
+Welcome to your new gem! In this directory, you'll find the files you need to be able to package up your Ruby library into a gem. Put your Ruby code in the file `lib/openscap_results_parser`. To experiment with that code, run `bin/console` for an interactive prompt.
 
 ## Installation
 
 Add this line to your application's Gemfile:
 
 ```ruby
-gem 'openscap_parser'
+gem 'openscap_results_parser'
 ```
 
 And then execute:
@@ -16,7 +16,7 @@ And then execute:
 
 Or install it yourself as:
 
-    $ gem install openscap_parser
+    $ gem install openscap_results_parser
 
 ## Usage
 
@@ -58,14 +58,14 @@ To install this gem onto your local machine, run `bundle exec rake install`. To 
 A Dockerfile is provided to allow a containerized development environment:
 
 ```
-docker build . -t openscap_parser # build the container image
-docker run -itv $PWD:/app:z openscap_parser rake # run tests
-docker run -itv $PWD:/app:z openscap_parser pry --gem # console
+docker build . -t openscap_results_parser # build the container image
+docker run -itv $PWD:/app:z openscap_results_parser rake # run tests
+docker run -itv $PWD:/app:z openscap_results_parser pry --gem # console
 ```
 
 ## Contributing
 
-Bug reports and pull requests are welcome on GitHub at https://github.com/elobato/openscap_parser. This project is intended to be a safe, welcoming space for collaboration, and contributors are expected to adhere to the [Contributor Covenant](http://contributor-covenant.org) code of conduct.
+Bug reports and pull requests are welcome on GitHub at https://github.com/elobato/openscap_results_parser. This project is intended to be a safe, welcoming space for collaboration, and contributors are expected to adhere to the [Contributor Covenant](http://contributor-covenant.org) code of conduct.
 
 ## License
 
@@ -73,4 +73,4 @@ The gem is available as open source under the terms of the [MIT License](https:/
 
 ## Code of Conduct
 
-Everyone interacting in the OpenscapParser project’s codebases, issue trackers, chat rooms and mailing lists is expected to follow the [code of conduct](https://github.com/elobato/openscap_parser/blob/master/CODE_OF_CONDUCT.md).
+Everyone interacting in the OpenscapParser project’s codebases, issue trackers, chat rooms and mailing lists is expected to follow the [code of conduct](https://github.com/elobato/openscap_results_parser/blob/master/CODE_OF_CONDUCT.md).
