--- lib/openscap_parser.rb
@@ -1,22 +1,0 @@
-# frozen_string_literal: true
-
-require 'openscap_parser/version'
-require 'openscap_parser/util'
-require 'openscap_parser/benchmarks'
-require 'openscap_parser/test_results'
-require 'openscap_parser/profiles'
-require 'openscap_parser/rules'
-require 'openscap_parser/rule_results'
-require 'openscap_parser/tailorings'
-
-require 'openscap_parser/xml_file'
-require 'openscap_parser/datastream_file'
-require 'openscap_parser/test_result_file'
-require 'openscap_parser/tailoring_file'
-
-require 'date'
-require 'railtie' if defined?(Rails)
-
-module OpenscapParser
-  class Error < StandardError; end
-end
