+++ lib/openscap_results_parser.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+require 'openscap_results_parser/version'
+require 'openscap_results_parser/util'
+require 'openscap_results_parser/benchmarks'
+require 'openscap_results_parser/test_results'
+require 'openscap_results_parser/profiles'
+require 'openscap_results_parser/rules'
+require 'openscap_results_parser/rule_results'
+require 'openscap_results_parser/tailorings'
+
+require 'openscap_results_parser/xml_file'
+require 'openscap_results_parser/datastream_file'
+require 'openscap_results_parser/test_result_file'
+require 'openscap_results_parser/tailoring_file'
+
+require 'date'
+require 'railtie' if defined?(Rails)
+
+module OpenscapParser
+  class Error < StandardError; end
+end
