+++ lib/openscap_results_parser/profile.rb
