+++ lib/openscap_results_parser/rule.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require 'openscap_parser/rule_identifier'
-require 'openscap_parser/rule_references'
-require 'openscap_parser/xml_file'
+require 'openscap_results_parser/rule_identifier'
+require 'openscap_results_parser/rule_references'
+require 'openscap_results_parser/xml_file'
 
 # Mimics openscap-ruby Rule interface
 module OpenscapParser
