+++ lib/openscap_results_parser/rule_identifier.rb
