+++ lib/openscap_results_parser/rule_reference.rb
