+++ lib/openscap_results_parser/rule_result.rb
