+++ lib/openscap_results_parser/tailoring.rb
