+++ lib/openscap_results_parser/tailoring_file.rb
