+++ lib/openscap_results_parser/util.rb
