+++ lib/openscap_results_parser/version.rb
