+++ lib/openscap_results_parser/xml_node.rb
