+++ openscap_results_parser.gemspec
@@ -1,17 +1,17 @@
 
 lib = File.expand_path("../lib", __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require "openscap_parser/version"
+require "openscap_results_parser/version"
 
 Gem::Specification.new do |spec|
-  spec.name          = 'openscap_parser'
+  spec.name          = 'openscap_results_parser'
   spec.version       = OpenscapParser::VERSION
   spec.authors       = ['Daniel Lobato Garcia', 'Andrew Kofink']
   spec.email         = ['me@daniellobato.me', 'ajkofink@gmail.com']
 
   spec.summary       = %q{Parse OpenSCAP content}
   spec.description   = %q{This gem is a Ruby interface into SCAP content. It can parse SCAP datastream files (i.e. ssg-rhel7-ds.xml), scan result files output by oscap eval, and tailoring files.}
-  spec.homepage      = 'https://github.com/dLobatog/openscap_parser'
+  spec.homepage      = 'https://github.com/dLobatog/openscap_results_parser'
   spec.license       = 'MIT'
 
   # Prevent pushing this gem to RubyGems.org. To allow pushes either set the 'allowed_push_host'
