+++ test/openscap_results_parser/test_result_file_test.rb
