+++ test/openscap_results_parser/xml_file_test.rb
