+++ test/test_helper.rb
@@ -1,5 +1,5 @@
 $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
-require "openscap_parser"
+require "openscap_results_parser"
 require 'pathname'
 
 require "minitest/autorun"
